Updates remove Manifest SHA from shared libraries when unneeded
