--- tests/jenkins/jobs/ArchiveAssembleUpload_Jenkinsfile_no_sha.txt
@@ -1,17 +1,0 @@
-   ArchiveAssembleUpload_Jenkinsfile.run()
-      ArchiveAssembleUpload_Jenkinsfile.pipeline(groovy.lang.Closure)
-         ArchiveAssembleUpload_Jenkinsfile.echo(Executing on agent [label:none])
-         ArchiveAssembleUpload_Jenkinsfile.stage(assemble from archive, groovy.lang.Closure)
-            ArchiveAssembleUpload_Jenkinsfile.script(groovy.lang.Closure)
-               ArchiveAssembleUpload_Jenkinsfile.archiveAssembleUpload({platform=linux, architecture=x64})
-                  archiveAssembleUpload.echo(Read BUILD_SHA_linux_x64_[SHA|LOCK|PATH].)
-                  archiveAssembleUpload.echo(sha.sha=null)
-                  archiveAssembleUpload.echo(sha.lock=null)
-                  archiveAssembleUpload.echo(sha.path=null)
-                  archiveAssembleUpload.legacySCM(groovy.lang.Closure)
-                  archiveAssembleUpload.library({identifier=jenkins@20211123, retriever=null})
-                  Messages.asBoolean()
-                  Messages.add(stage, Skipped stage, linux-x64 was not built.)
-                     archiveAssembleUpload.writeFile({file=messages/stage.msg, text=Skipped stage, linux-x64 was not built.})
-                     archiveAssembleUpload.stash({includes=messages/*, name=messages-stage})
-                  archiveAssembleUpload.echo(Skipping, linux x64 was not built.)
