+++ tests/jenkins/jobs/AssembleUpload_Jenkinsfile
@@ -6,11 +6,6 @@ pipeline {
                 script {
                     assembleUpload(
                         manifest: 'tests/data/opensearch-build-1.1.0.yml',
-                        sha: [
-                            sha: 'sha1',
-                            lock: 'tests/jenkins/data/opensearch-1.3.0.yml.lock',
-                            path: 'assemble-upload-build/1.1.0/shas/linux/x64/sha1.yml'
-                        ]
                     )
                 }
             }
