+++ tests/jenkins/jobs/BuildAndUploadManifestSHA_Jenkinsfile
@@ -1,10 +1,10 @@
 pipeline {
     agent none
     stages {
-        stage('assemble from archive') {
+        stage('upload manifest SHA') {
             steps {
                 script {
-                    archiveAssembleUpload(
+                    buildAndUploadManifestSHA(
                         manifest: 'tests/jenkins/data/opensearch-1.3.0.yml',
                         platform: 'linux',
                         architecture: 'x64'
