+++ tests/jenkins/jobs/BuildArchive_Jenkinsfile.txt
@@ -1,0 +1,17 @@
+   BuildArchive_Jenkinsfile.run()
+      BuildArchive_Jenkinsfile.pipeline(groovy.lang.Closure)
+         BuildArchive_Jenkinsfile.echo(Executing on agent [label:none])
+         BuildArchive_Jenkinsfile.stage(build and archive, groovy.lang.Closure)
+            BuildArchive_Jenkinsfile.script(groovy.lang.Closure)
+               BuildArchive_Jenkinsfile.buildArchive({manifest=tests/jenkins/data/opensearch-1.3.0.yml, platform=linux, architecture=x64})
+                  buildArchive.legacySCM(groovy.lang.Closure)
+                  buildArchive.library({identifier=jenkins@20211123, retriever=null})
+                  buildArchive.buildManifest({manifest=tests/jenkins/data/opensearch-1.3.0.yml, platform=linux, architecture=x64})
+                     buildManifest.sh(./build.sh tests/jenkins/data/opensearch-1.3.0.yml -p linux -a x64)
+                  buildArchive.echo(Archiving into zip: builds/**, tests/jenkins/data/opensearch-1.3.0.yml)
+                  Messages.asBoolean()
+                  Messages.add(stage, Built stage.)
+                     buildArchive.writeFile({file=messages/stage.msg, text=Built stage.})
+                     buildArchive.stash({includes=messages/*, name=messages-stage})
+                  buildArchive.readYaml({file=tests/jenkins/data/opensearch-1.3.0.yml})
+                  InputManifest.asBoolean()
