+++ tests/jenkins/jobs/BuildAssembleUpload_Jenkinsfile.txt
@@ -6,36 +6,21 @@
                BuildAssembleUpload_Jenkinsfile.buildAssembleUpload({manifest=tests/jenkins/data/opensearch-1.3.0.yml, platform=linux, architecture=x64})
                   buildAssembleUpload.legacySCM(groovy.lang.Closure)
                   buildAssembleUpload.library({identifier=jenkins@20211123, retriever=null})
-                  buildAssembleUpload.getManifestSHA({manifest=tests/jenkins/data/opensearch-1.3.0.yml, platform=linux, architecture=x64})
-                     getManifestSHA.buildManifest({manifest=tests/jenkins/data/opensearch-1.3.0.yml, platform=linux, architecture=x64, lock=true})
-                        buildManifest.sh(./build.sh tests/jenkins/data/opensearch-1.3.0.yml -p linux -a x64 --lock)
-                     getManifestSHA.sha1(tests/jenkins/data/opensearch-1.3.0.yml.lock)
-                     getManifestSHA.echo(Manifest SHA: sha1)
-                     getManifestSHA.legacySCM(groovy.lang.Closure)
-                     getManifestSHA.library({identifier=jenkins@20211123, retriever=null})
-                     getManifestSHA.readYaml({file=tests/jenkins/data/opensearch-1.3.0.yml.lock})
-                     InputManifest.asBoolean()
-                     InputManifest.getSHAsRoot(vars-build)
-                     getManifestSHA.echo(Manifest lock: tests/jenkins/data/opensearch-1.3.0.yml.lock)
-                     getManifestSHA.echo(Manifest SHA path: vars-build/1.3.0/shas/sha1.yml)
-                     getManifestSHA.withAWS({role=opensearch-bundle, roleAccount=account, duration=900, roleSessionName=jenkins-session}, groovy.lang.Closure)
-                        getManifestSHA.s3DoesObjectExist({bucket=artifact-bucket, path=vars-build/1.3.0/shas/sha1.yml})
-                     getManifestSHA.echo(Manifest SHA exists: true)
-                  buildAssembleUpload.readYaml({file=tests/jenkins/data/opensearch-1.3.0.yml.lock})
+                  buildAssembleUpload.readYaml({file=tests/jenkins/data/opensearch-1.3.0.yml})
                   InputManifest.asBoolean()
-                  buildAssembleUpload.buildManifest({manifest=tests/jenkins/data/opensearch-1.3.0.yml.lock, platform=linux, architecture=x64})
-                     buildManifest.sh(./build.sh tests/jenkins/data/opensearch-1.3.0.yml.lock -p linux -a x64)
-                  buildAssembleUpload.assembleUpload({manifest=builds/opensearch/manifest.yml, platform=linux, architecture=x64, sha={sha=sha1, lock=tests/jenkins/data/opensearch-1.3.0.yml.lock, path=vars-build/1.3.0/shas/sha1.yml, exists=true}})
+                  buildAssembleUpload.buildManifest({manifest=tests/jenkins/data/opensearch-1.3.0.yml, platform=linux, architecture=x64})
+                     buildManifest.sh(./build.sh tests/jenkins/data/opensearch-1.3.0.yml -p linux -a x64)
+                  buildAssembleUpload.assembleUpload({manifest=builds/opensearch/manifest.yml, platform=linux, architecture=x64})
                      assembleUpload.legacySCM(groovy.lang.Closure)
                      assembleUpload.library({identifier=jenkins@20211123, retriever=null})
-                     assembleUpload.assembleManifest({manifest=builds/opensearch/manifest.yml, platform=linux, architecture=x64, sha={sha=sha1, lock=tests/jenkins/data/opensearch-1.3.0.yml.lock, path=vars-build/1.3.0/shas/sha1.yml, exists=true}})
+                     assembleUpload.assembleManifest({manifest=builds/opensearch/manifest.yml, platform=linux, architecture=x64})
                         assembleManifest.legacySCM(groovy.lang.Closure)
                         assembleManifest.library({identifier=jenkins@20211123, retriever=null})
                         assembleManifest.readYaml({file=builds/opensearch/manifest.yml})
                         BuildManifest.asBoolean()
                         BuildManifest.getArtifactRootUrl(https://ci.opensearch.org/dbc, vars-build, 33)
                         assembleManifest.sh(./assemble.sh "builds/opensearch/manifest.yml" --base-url https://ci.opensearch.org/dbc/vars-build/1.1.0/33/linux/x64)
-                     assembleUpload.uploadArtifacts({manifest=builds/opensearch/manifest.yml, platform=linux, architecture=x64, sha={sha=sha1, lock=tests/jenkins/data/opensearch-1.3.0.yml.lock, path=vars-build/1.3.0/shas/sha1.yml, exists=true}})
+                     assembleUpload.uploadArtifacts({manifest=builds/opensearch/manifest.yml, platform=linux, architecture=x64})
                         uploadArtifacts.legacySCM(groovy.lang.Closure)
                         uploadArtifacts.library({identifier=jenkins@20211123, retriever=null})
                         uploadArtifacts.readYaml({file=builds/opensearch/manifest.yml})
@@ -52,11 +37,9 @@ https://ci.opensearch.org/dbc/vars-build/1.1.0/33/linux/x64/dist/opensearch/mani
                            uploadArtifacts.writeFile({file=messages/stage.msg, text=https://ci.opensearch.org/dbc/vars-build/1.1.0/33/linux/x64/builds/opensearch/manifest.yml
 https://ci.opensearch.org/dbc/vars-build/1.1.0/33/linux/x64/dist/opensearch/manifest.yml})
                            uploadArtifacts.stash({includes=messages/*, name=messages-stage})
-                     assembleUpload.withAWS({role=opensearch-bundle, roleAccount=account, duration=900, roleSessionName=jenkins-session}, groovy.lang.Closure)
-                        assembleUpload.s3Upload({bucket=artifact-bucket, file=tests/jenkins/data/opensearch-1.3.0.yml.lock, path=vars-build/1.3.0/shas/sha1.yml})
-                     assembleUpload.readYaml({file=tests/jenkins/data/opensearch-1.3.0.yml.lock})
+                     assembleUpload.readYaml({file=builds/opensearch/manifest.yml})
                      InputManifest.asBoolean()
                      assembleUpload.readYaml({file=builds/opensearch/manifest.yml})
                      BuildManifest.asBoolean()
                      InputManifest.getPublicDistUrl(https://ci.opensearch.org/dbc, vars-build, 33, linux, x64)
-                     assembleUpload.echo(Setting env."ARTIFACT_URL_linux_x64"=https://ci.opensearch.org/dbc/vars-build/1.3.0/33/linux/x64/dist/opensearch/opensearch-1.3.0-linux-x64.tar.gz)
+                     assembleUpload.echo(Setting env."ARTIFACT_URL_linux_x64"=https://ci.opensearch.org/dbc/vars-build/1.1.0/33/linux/x64/dist/opensearch/opensearch-1.1.0-linux-x64.tar.gz)
